Dermo_AI/requirements.txt
# Core ML Libraries (CPU-only versions)
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cpu
transformers==4.35.0
sentence-transformers==2.2.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Image Processing
pillow==10.1.0
opencv-python-headless==4.8.1.78  # Lighter version without GUI deps

# Core Utilities
numpy==1.26.0
streamlit==1.32.0

# Optional (for Hugging Face Spaces)
huggingface-hub==0.20.2